+++ packages/firebase_auth/firebase_auth/lib/src/auth_provider/twitter_auth_provider.dart
@@ -1,0 +1,19 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+part of firebase_auth;
+
+class TwitterAuthProvider {
+  static const String providerId = 'twitter.com';
+
+  static AuthCredential getCredential({
+    @required String authToken,
+    @required String authTokenSecret,
+  }) {
+    return TwitterAuthCredential(
+      authToken: authToken,
+      authTokenSecret: authTokenSecret,
+    );
+  }
+}