+++ packages/firebase_in_app_messaging/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.1
+
+* Updated the plugin API to v2.
+
+## 0.1.0
+
+* Updated README instructions for contributing for consistency with other Flutterfire plugins.
+
 ## 0.0.1+4
 
 * Update example app with correct const constructors.